Clean up device.py
