+++ broadlink/device.py
@@ -4,84 +4,12 @@ import socket
 import threading
 import random
 import time
-from typing import Generator, Tuple, Union
+import typing as t
 
 from . import exceptions as e
-from .protocol import Datetime, BlockSizeHandler, EncryptionHandler, ExtBlockSizeHandler
+from . import protocol as p
 
 _LOGGER = logging.getLogger(__name__)
-
-HelloResponse = Tuple[int, Tuple[str, int], str, str, bool]
-
-
-def scan(
-    timeout: int = 10,
-    local_ip_address: str = None,
-    discover_ip_address: str = "255.255.255.255",
-    discover_ip_port: int = 80,
-) -> Generator[HelloResponse, None, None]:
-    """Broadcast a hello message and yield responses."""
-    conn = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
-    conn.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
-    conn.setsockopt(socket.SOL_SOCKET, socket.SO_BROADCAST, 1)
-
-    if local_ip_address:
-        conn.bind((local_ip_address, 0))
-        port = conn.getsockname()[1]
-    else:
-        local_ip_address = "0.0.0.0"
-        port = 0
-
-    packet = bytearray(0x30)
-    packet[0x08:0x14] = Datetime.pack(Datetime.now())
-    packet[0x18:0x1C] = socket.inet_aton(local_ip_address)[::-1]
-    packet[0x1C:0x1E] = port.to_bytes(2, "little")
-    packet[0x26] = 6
-
-    checksum = sum(packet, 0xBEAF) & 0xFFFF
-    packet[0x20:0x22] = checksum.to_bytes(2, "little")
-
-    start_time = time.time()
-    discovered = []
-
-    try:
-        while (time.time() - start_time) < timeout:
-            time_left = timeout - (time.time() - start_time)
-            conn.settimeout(min(1, time_left))
-            conn.sendto(packet, (discover_ip_address, discover_ip_port))
-
-            while True:
-                try:
-                    resp, host = conn.recvfrom(1024)
-                except socket.timeout:
-                    break
-
-                devtype = resp[0x34] | resp[0x35] << 8
-                mac = resp[0x3A:0x40][::-1]
-
-                if (host, mac, devtype) in discovered:
-                    continue
-                discovered.append((host, mac, devtype))
-
-                name = resp[0x40:].split(b"\x00")[0].decode()
-                is_locked = bool(resp[-1])
-                yield devtype, host, mac, name, is_locked
-    finally:
-        conn.close()
-
-
-def ping(address: str, port: int = 80) -> None:
-    """Send a ping packet to an address.
-
-    This packet feeds the watchdog timer of firmwares >= v53.
-    Useful to prevent reboots when the cloud cannot be reached.
-    It must be sent every 2 minutes in such cases.
-    """
-    with socket.socket(socket.AF_INET, socket.SOCK_DGRAM) as conn:
-        conn.setsockopt(socket.SOL_SOCKET, socket.SO_BROADCAST, 1)
-        packet = bytearray(0x30)
-        packet[0x26] = 1
-        conn.sendto(packet, (address, port))
 
 
 class device:
@@ -89,13 +17,13 @@ class device:
 
     TYPE = "Unknown"
 
-    _BS_HDLR = BlockSizeHandler
-    _ENC_HDLR = EncryptionHandler
+    BlockSizeHandler = p.BlockSizeHandler
+    EncryptionHandler = p.EncryptionHandler
 
     def __init__(
         self,
-        host: Tuple[str, int],
-        mac: Union[bytes, str] = None,
+        host: t.Tuple[str, int],
+        mac: t.Union[bytes, str] = None,
         devtype: int = None,
         timeout: int = 10,
         name: str = None,
@@ -112,13 +40,12 @@ class device:
         self.model = model
         self.manufacturer = manufacturer
         self.is_locked = is_locked
+        self.type = self.TYPE  # For backwards compatibility.
 
         self._count = random.randint(0x8000, 0xFFFF)
-        self._bs_hdlr = self._BS_HDLR()
-        self._enc_hdlr = self._ENC_HDLR()
+        self._bs_hdlr = self.BlockSizeHandler()
+        self._enc_hdlr = self.EncryptionHandler()
         self._lock = threading.Lock()
-
-        self.type = self.TYPE  # For backwards compatibility.
 
     def __repr__(self) -> str:
         """Return a formal representation of the device."""
@@ -155,27 +82,14 @@ class device:
         info = " / ".join(info)
         return "%s (%s)" % (self.name or "Unknown", info)
 
-    def auth(self) -> bool:
-        """Authenticate to the device.
+    def ping(self) -> None:
+        """Send a Ping Response packet to the device.
 
-        Send a Client Key Exchange packet to the device and update
-        device control ID and AES key with the response.
+        This packet feeds the watchdog timer of firmwares >= v53.
+        Useful to prevent reboots when the cloud cannot be reached.
+        It must be sent every 2 minutes in such cases.
         """
-        with self._lock:
-            self._enc_hdlr.reset()
-            payload = bytearray(0x50)
-            payload[0x04:0x13] = [0x31] * 15
-            payload[0x1E] = 0x01
-            payload[0x2D] = 0x01
-            payload[0x30:0x36] = "Test 1".encode()
-            resp, err = self.send_packet(0x65, payload)
-            if err:
-                raise e.exception(err)
-
-            conn_id = int.from_bytes(resp[:0x4], "little")
-            key = resp[0x04:0x14]
-            self._enc_hdlr.update(conn_id, key)
-            return True
+        self.send_packet(0x01)
 
     def hello(self) -> None:
         """Start communicating with the device.
@@ -219,48 +133,27 @@ class device:
         self.name = resp[0x10:].split(b"\x00")[0].decode()
         self.is_locked = bool(resp[-1])
 
-    def ping(self) -> None:
-        """Ping the device.
+    def auth(self) -> bool:
+        """Authenticate to the device.
 
-        This packet feeds the watchdog timer of firmwares >= v53.
-        Useful to prevent reboots when the cloud cannot be reached.
-        It must be sent every 2 minutes in such cases.
+        Send a Client Key Exchange packet to the device and update
+        device control ID and AES key with the response.
         """
-        ping(self.host[0], port=self.host[1])
+        with self._lock:
+            self._enc_hdlr.reset()
+            payload = bytearray(0x50)
+            payload[0x04:0x13] = [0x31] * 15
+            payload[0x1E] = 0x01
+            payload[0x2D] = 0x01
+            payload[0x30:0x36] = "Test 1".encode()
+            resp, err = self.send_packet(0x65, payload)
+            if err:
+                raise e.exception(err)
 
-    def get_fwversion(self) -> int:
-        """Get firmware version."""
-        packet = bytearray([0x68])
-        resp, err = self.send_packet(0x6A, packet)
-        if err:
-            raise e.exception(err)
-        return resp[0x04] | resp[0x05] << 8
-
-    def set_name(self, name: str) -> None:
-        """Set device name."""
-        packet = bytearray(4)
-        packet += name.encode("utf-8")
-        packet += bytearray(0x50 - len(packet))
-        packet[0x43] = bool(self.is_locked)
-        err = self.send_packet(0x6A, packet)[1]
-        if err:
-            raise e.exception(err)
-        self.name = name
-
-    def set_lock(self, state: bool) -> None:
-        """Lock/unlock the device."""
-        packet = bytearray(4)
-        packet += self.name.encode("utf-8")
-        packet += bytearray(0x50 - len(packet))
-        packet[0x43] = bool(state)
-        err = self.send_packet(0x6A, packet)[1]
-        if err:
-            raise e.exception(err)
-        self.is_locked = bool(state)
-
-    def get_type(self) -> str:
-        """Return device type."""
-        return self.type
+            conn_id = int.from_bytes(resp[:0x4], "little")
+            key = resp[0x04:0x14]
+            self._enc_hdlr.update(conn_id, key)
+            return True
 
     def send_cmd(
         self,
@@ -276,12 +169,47 @@ class device:
             raise e.exception(err)
         return self._bs_hdlr.unpack(resp)[0x4:]
 
+    def set_name(self, name: str) -> None:
+        """Set device name."""
+        cname = name.encode()
+        packet = bytearray(0x50)
+        packet[0x04:0x04+len(cname)] = cname
+        packet[0x43] = bool(self.is_locked)
+        err = self.send_packet(0x6A, packet)[1]
+        if err:
+            raise e.exception(err)
+        self.name = name
+
+    def set_lock(self, state: bool) -> None:
+        """Lock/unlock the device."""
+        name = self.name.encode()
+        packet = bytearray(0x50)
+        packet[0x04:0x04+len(name)] = name
+        packet[0x43] = bool(state)
+        err = self.send_packet(0x6A, packet)[1]
+        if err:
+            raise e.exception(err)
+        self.is_locked = bool(state)
+
+    def get_fwversion(self) -> int:
+        """Get firmware version."""
+        packet = bytearray(0x10)
+        packet[0] = 0x68
+        resp, err = self.send_packet(0x6A, packet)
+        if err:
+            raise e.exception(err)
+        return resp[0x04] | resp[0x05] << 8
+
+    def get_type(self) -> str:
+        """Return device type."""
+        return self.type
+
     def send_packet(
         self,
         info_type: int,
         payload: bytes = b"",
         retry_intvl: float = 1.0,
-    ) -> bytes:
+    ) -> t.Union[bytes, None]:
         """Send a packet to the device."""
         packet = bytearray(0x30)
         payload = bytes(payload)
@@ -308,12 +236,15 @@ class device:
         # Public request.
         elif info_type >> 5 == 0 and not info_type % 2:
             exp_resp_type = info_type + 1
-
-            packet = bytearray(0x30)
-            packet[0x08:0x14] = Datetime.pack(Datetime.now())
+            packet[0x08:0x14] = p.Datetime.pack(p.Datetime.now())
             # packet[0x18:0x1E] = Address.pack((local_ip_addr, port))
             packet[0x26:0x28] = info_type.to_bytes(2, "little")
             packet.extend(payload)
+
+        # Ping response.
+        elif info_type == 1:
+            exp_resp_type = None
+            packet[0x26:0x28] = info_type.to_bytes(2, "little")
 
         # Encrypted response / public response.
         else:
@@ -350,6 +281,9 @@ class device:
                 conn.sendto(packet, self.host)
 
                 _LOGGER.debug("%s sent to %s", packet, self.host)
+
+                if exp_resp_type is None:
+                    return
 
                 try:
                     resp, src = conn.recvfrom(2048)
@@ -421,5 +355,5 @@ class v4(type):
     def __new__(cls, name, bases, dct):
         """Create a new device."""
         dev_cls = super().__new__(cls, name, bases, dct)
-        dev_cls._BS_HDLR = ExtBlockSizeHandler
+        dev_cls.BlockSizeHandler = p.ExtBlockSizeHandler
         return dev_cls
